struct Rectangle

+++ Project1/Vector2D.cpp
@@ -6,7 +6,7 @@ Vector2D::Vector2D():
 {
 }
 
-Vector2D::Vector2D(float x, float y):
+Vector2D::Vector2D(int x, int y):
 	x(x),
 	y(y)
 {
